Refactored to flatten board state into a hash-map
The `doall`s in the board render method are ugly, but reagent
won't pick up on the ratom being deferenced in a component inside
of a lazy sequence, like the one produced by `for`. If the dereference
isn't picked up the cell won't rerender on click.
On the engine side thnings are much cleaner. For instance there's no
longer a need for an `update-board-state-function` because there is
really only cell state, and that's all there ever was!